Introduce the draggable titlebar, match color (for dark only for now), move menu items around
