+++ resources/js/Layouts/Navigation/components/NavigationBar.jsx
@@ -12,11 +12,12 @@ import useMediaQuery from '../../../hooks/useMediaQuery';
 import useWidth from '../../../hooks/useWidth';
 import clsx from 'clsx';
 
-export default function NavigationBar({ auth, main, overflow }) {
+export default function NavigationBar({ main, overflow, socials }) {
 	const matchesSmUp = useMediaQuery('(min-width: 576px)');
 	const ref = useRef(null);
 	const width = useWidth(ref);
-    const page = usePage();
+	const page = usePage();
+  const { auth } = usePage().props;
 
 	const [mainItems, overflowItems] = useMemo(() => {
 		const maxVisibleItems = Math.floor(width / 65);
@@ -44,11 +45,6 @@ export default function NavigationBar({ auth, main, overflow }) {
 		return !!overflowUrls.find((url) => page.url.includes(url));
 	}, [overflowItems]);
 
-    function handleLogout(e) {
-        e.preventDefault();
-        Inertia.post('/logout');
-    }
-
 	return (
 		<div className="sidebar" ref={ref}>
 			{mainItems.map((item, key) => {
@@ -75,7 +71,7 @@ export default function NavigationBar({ auth, main, overflow }) {
 
 					<ul className="dropdown-menu">
 						{overflowItems.map((item, key) => {
-    						const Component = item.type === 'external' ? 'a' : InertiaLink;
+							const Component = item.type === 'external' ? 'a' : InertiaLink;
 							const mainProps = item.type === 'external' ? { target: '_blank' } : {};
 
 							if (item.type === 'link') {
@@ -84,7 +80,7 @@ export default function NavigationBar({ auth, main, overflow }) {
 										{...mainProps}
 										key={key}
 										href={`${item.url}${item.primary ?? ''}`}
-										className={clsx('dropdown-item', { 'active': page.url.includes(item.url)})}
+										className={clsx('dropdown-item', { 'active': page.url.includes(item.url) })}
 									>
 										<AmaranthIcon icon={item.icon} /> {item.title}
 									</Component>
@@ -95,7 +91,7 @@ export default function NavigationBar({ auth, main, overflow }) {
 										{...mainProps}
 										key={key}
 										href={`${item.url}${item.primary ?? ''}`}
-										className={clsx('dropdown-item', { 'active': page.url.includes(item.url)})}
+										className={clsx('dropdown-item', { 'active': page.url.includes(item.url) })}
 									>
 										<AmaranthIcon icon={item.icon} /> {item.title}
 									</Component>
@@ -104,28 +100,6 @@ export default function NavigationBar({ auth, main, overflow }) {
 								return (<div className="dropdown-divider" key={key} />);
 							}
 						})}
-						<div className="dropdown-divider" />
-						{auth ?
-							<>
-								<InertiaLink href="/settings" className={clsx('dropdown-item', { 'active': page.url.includes('/settings')})}>
-									<AmaranthIcon icon={aiGear} /> Settings
-								</InertiaLink>
-								<form onSubmit={handleLogout}>
-									<button type="submit" className="dropdown-item">
-										<AmaranthIcon icon={aiArrowRightFromBracket} /> Log out
-									</button>
-								</form>
-							</>
-						:
-							<>
-								<InertiaLink href="/settings" className="dropdown-item">
-									<AmaranthIcon icon={aiGear} /> Settings
-								</InertiaLink>
-								<InertiaLink href="/login" className="dropdown-item">
-									<AmaranthIcon icon={aiArrowRightToBracket} /> Sign in
-								</InertiaLink>
-							</>
-						}
 					</ul>
 				</>
 			}
@@ -133,19 +107,9 @@ export default function NavigationBar({ auth, main, overflow }) {
 			{matchesSmUp &&
 				<>
 					<div className="flex-grow-1 d-none d-sm-block" />
-					<NavigationItem url="/settings" icon={aiGear} title="Settings" />
-
-					{auth ?
-						<form onSubmit={handleLogout} className="d-none d-sm-block">
-							<button type="submit" className="sidebar-item">
-								<AmaranthIcon icon={aiArrowRightFromBracket} /> <span className="sidebar-label">Log out</span>
-							</button>
-						</form>
-					:
-						<InertiaLink href="/login" className="sidebar-item">
-							<AmaranthIcon icon={aiArrowRightToBracket} /> <span className="sidebar-label">Sign in</span>
-						</InertiaLink>
-					}
+					{socials.map((item, key) => (
+						<NavigationItem url={item.url} icon={item.icon} primary={item.primary} title={item.title} key={key} external />
+					))}
 				</>
 			}
 		</div>
