+++ resources/sass/_navbar.scss
@@ -2,7 +2,15 @@
     height: 48px;
     padding: 0;
 
+    &.navbar-main {
+        left: env(titlebar-area-x, 0);
+        width: env(titlebar-area-width, 100%);
+        transition: $transition-base;
+    }
+
     .container-fluid {
+        flex-direction: row;
+        justify-content: space-between;
     }
     
     .navbar-brand {
