+++ resources/views/app.blade.php
@@ -13,7 +13,6 @@
         
         <link rel="shortcut icon" href="{{ env('APP_PREVIEW') === 'preview' ? asset('images/logo-preview.svg') : (env('APP_PREVIEW') === 'canary' ? asset('images/logo-canary.svg') : asset('images/logo.svg')) }}">
         <link rel="manifest" href="{{ env('APP_PREVIEW') === 'preview' ? asset('manifest-preview.json') : (env('APP_PREVIEW') === 'canary' ? asset('manifest-canary.json') : asset('manifest.json')) }}">
-        <meta name="theme-color" content="#0066ff">
         <link rel="apple-touch-icon" href="{{ env('APP_PREVIEW') === 'preview' ? asset('images/logo-preview-mask-192.png') : (env('APP_PREVIEW') === 'canary' ? asset('images/logo-canary-mask-192.png') : asset('images/logo-mask-192.png')) }}">
     </head>
     <body>
